Statistic CFD investigation in aneurysms
Study on the relationship between wall shear stress and aspect ratio of cerebral aneurysms with different pressure differences using CFD simulations.
CFD simulations were performed for 60 human cerebral aneurysms (30 previously ruptured and 30 previously unruptured) to study the behavior of the time-averaged wall shear stress (TAWSS) with respect to the aspect ratio (AR), implementing a set of low, normal, and high-pressure differences between the inlet and the outlets of each artery. It is well known that exists a direct relationship between TAWSS and the rupture. In this investigation, we presented an important result because the condition of the pressure among branches and the AR may be measured in any patient, then a slope may be associated, and finally a TAWSS may be estimated. We found that whether the pressure difference increased, the absolute slopes between TAWSS and AR increased as well. Also, the magnitude of the slope in the previously unruptured aneurysms was 4.7 times the slope in the previously ruptured aneurysms. On the other hand, TAWSS was higher in the previously unruptured aneurysm than previously ruptured aneurysms due to unruptured aneurysms have a smaller surface area. Furthermore, we analyzed the relationship between TAWSS and other geometric parameters of the aneurysm, such as bottleneck and non-sphericity index; however, no correlation was found for either cases.